Test.java
Develop a Java program for a professional training company to read the names, grades and attendance of students in a given course.
It is known that approval is required for a score equal to or greater than 70 points and a frequency equal to or greater than 80%. In addition, due to an operating rule of the company,
 classes must have a fixed number of 10 students per class.

Attention: The specifications and steps for the elaboration of the program must be followed in order to avoid errors in the code and the penalty of the notes in each one of the
requested items.

The complete code in Java must be posted, and each of the characteristics will be evaluated accordingly. Therefore, the program should:



(note: 10%) Fill directly in the code a vector of Strings with the name of 10 students;
(grade: 20%) Read via console as grades (0-100) and frequencies (0-100) of each of the 10 students showing the names of each one at the time of reading the values.

For example:

Student: João da Silva

Enter the score (0-100):

Enter the frequency (0-100):

 

Store as student notes in a Double array;

Store as student frequencies in a vector of integers;

(grade: 10%) Present the names of all motivated students, their grades and frequency;
(grade: 10%) Presenting the names of all failed students, their grades and their frequency;
(grade: 10%) Presenting the name of the student who presented the best grade and his grade;
(grade: 10%) Shows the name of the student who had the worst grade and his grade;
(note: 10%) Present the percentage of students initiated in the class;
(grade: 10%) Presentation of the percentage of students failing the class;
(grade: 10%) Presenting an average of the class grades.